+++ ptn/missionalertbot/_metadata.py
@@ -1,2 +1,2 @@
 # version is BREAKING CHANGE - MAJOR CHANGE - MINOR CHANGE
-__version__ = '2.3.1'
+__version__ = '2.3.2'